frontend/requirements.txt
streamlit==1.28.2
requests==2.31.0
python-dotenv==1.0.0
streamlit-chat==0.1.1